Refactor: AnswerTimeProgress 컴포넌트 코드 리팩토링
