+++ src/container/interviewVideo/components/answerTimeProgressGroup/components/answerTimeProgress/index.tsx
@@ -1,9 +1,5 @@
 'use client';
 
-import { useEffect } from 'react';
-
-import useInterviewProgress from '@/stores/interviewProgress';
-
 import useAnswerTimer from './hooks/useAnswerTimer';
 import { IProps } from './types';
 
@@ -13,17 +9,14 @@ const AnswerTimeProgress = ({
   enabled,
   onEnded,
 }: IProps) => {
-  const { time } = useAnswerTimer({ defaultTime, timer, enabled, onEnded });
-  const setInterview = useInterviewProgress((state) => state.setInterview);
-  const progressWidth = (time / timer) * 100;
+  const { progressWidth } = useAnswerTimer({
+    defaultTime,
+    timer,
+    enabled,
+    onEnded,
+  });
   const commonStyle = 'rounded-2xl bg-primaries-primary border';
   const animationStyle = `${enabled ? 'animate-pulse' : ''}`;
-
-  useEffect(() => {
-    if (enabled) {
-      setInterview({ progressingTime: time });
-    }
-  }, [enabled, setInterview, time]);
 
   return (
     <li className={`h-full w-full bg-opacity-20 ${commonStyle}`}>
